Check if ssh-agent is running before killall
If it's not running killall returns 1 and causes failure exit due to set
-e.

+++ assets/helpers/utils.sh
@@ -167,7 +167,9 @@ getBasePathOfBitbucket() {
 
 cleanup() {
   rm -rf "$TMPDIR/bitbucket-pullrequest-resource-*"
-  killall ssh-agent
+  if pgrep ssh-agent; then
+    killall ssh-agent
+  fi
 }
 
 trap cleanup EXIT
